requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.2
Flask==2.2.2
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.2
mistune==2.0.4
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.0
pandocfilters==1.5.0
platformdirs==2.6.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.2.2
